girona-back/requirements.txt
fastapi==0.127.0
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.45
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic-settings==2.10.1
email-validator==2.2.0
python-jose==3.5.0
passlib[bcrypt]==1.7.4
openpyxl==3.1.5
google-api-python-client
google-auth
